Refactor id as Long instead of int

+++ src/main/java/com/cottonfactory/products/repositories/JacketRepository.java
@@ -3,5 +3,5 @@ package com.cottonfactory.products.repositories;
 import com.cottonfactory.products.entities.Jacket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JacketRepository extends JpaRepository<Jacket , Integer> {
+public interface JacketRepository extends JpaRepository<Jacket , Long> {
 }
